refactor: 삭제 핸들러 분리
