+++ src/components/pages/CartPage/CartListSection/CartListSection.tsx
@@ -13,7 +13,7 @@ import {
 import * as Text from '@commons/Text/Text';
 import CardList from '@components/pages/CartPage/CartListSection/CartList/CartList';
 import { Checkbox as WholeCheckbox } from '@components/commons/Checkbox/Checkbox';
-import { Button as DeleteButton } from '@components/commons/Button/Button';
+import { Button as CheckedItemsDeleteButton } from '@components/commons/Button/Button';
 
 const CartListSection = () => {
   const cartItemLength = useRecoilValue(cartItemsLengthSelector);
@@ -24,23 +24,10 @@ const CartListSection = () => {
   const setCartItems = useSetRecoilState(cartItemsState);
   const [isDeleteItem, setIsDeleteItem] = useState(false);
 
-  const handleDeleteButtonClick = (productId?: number) => {
-    if (productId) {
-      deleteData(`/${productId}`);
-      setCartItems(prev => {
-        const newCartItems = { ...prev };
-        const key = `product${productId}`;
-        delete newCartItems[key];
-
-        return newCartItems;
-      });
-      setIsDeleteItem(() => true);
-
-      return;
-    }
-
+  const handleCheckedItemsDeleteButtonClick = () => {
     const productIds = Object.values(checkedCartItems);
     const urls = productIds.map(productId => `/${productId}`);
+
     productIds.forEach((productId: number, index) => {
       deleteData(urls[index]);
       setCartItems(prev => {
@@ -51,7 +38,9 @@ const CartListSection = () => {
         return newCartItems;
       });
     });
+
     setIsDeleteItem(() => true);
+
     setCheckedCartItems(prev => {
       const newCheckedCartItems = {
         ...prev,
@@ -74,26 +63,22 @@ const CartListSection = () => {
       <StyledCartListTextBox>
         <Text.Paragraph>든든배송 상품 ({cartItemLength}개)</Text.Paragraph>
       </StyledCartListTextBox>
-      <CardList
-        isDeleteItem={isDeleteItem}
-        setIsDeleteItem={setIsDeleteItem}
-        handleDeleteButtonClick={handleDeleteButtonClick}
-      />
+      <CardList isDeleteItem={isDeleteItem} setIsDeleteItem={setIsDeleteItem} />
       <StyledCartListFlexBox>
         <WholeCheckbox />
         <Text.Description>
           전체선택 ({Object.keys(checkedCartItems).length}/{cartItemLength})
         </Text.Description>
-        <DeleteButton
+        <CheckedItemsDeleteButton
           width="100px"
           height="36px"
           padding="4px"
           border="1px solid #BBBBBB"
           border-radius="0px"
-          onClick={() => handleDeleteButtonClick()}
+          onClick={() => handleCheckedItemsDeleteButtonClick()}
         >
           선택삭제
-        </DeleteButton>
+        </CheckedItemsDeleteButton>
       </StyledCartListFlexBox>
     </StyledCartListSection>
   );
